Dashboard page divided into components...
